Rename to issuefilter + add duplicate detection
